vrag/requirements.txt
pdf2image
qwen-vl-utils
byaldi
torch
pypdf
tqdm
pillow
colpali-engine
pyvespa
requests
numpy
python-dotenv
PyPDF2
